requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-telegram-bot==21.6
python-dotenv==1.0.1
python-multipart==0.0.9
supabase==2.28.0
